+++ CHANGELOG.rst
@@ -2,7 +2,7 @@
 ------------------------
 * Change: allow getting all subroles for a specific role in _all_roles_list
   [tiberich #21218]
-* Feature: allow excluding disabled users in search_user
+* Feature: allow excluding disabled users in search_user and search_user_email
   [tiberich #22488]
 
 1.3.23 (2014-10-02)